fix: dockerFile and config.sh

+++ start.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
-PLAY_DIR=/home/ubuntu/taitra-1.0
-CONFIG_FILE=$PLAY_DIR/config/application.conf
+PLAY_DIR=taitra-1.0
+CONFIG_FILE=$PLAY_DIR/conf/application.conf
 
 
 
-if [-n "$SECRET_KEY"] && [ -n "$DB_URL" ] && [ -n "$DB_USERNAME" ] && [ -n "$DB_PASSWORD" ]; then
+if [ -n "$SECRET_KEY" ] && [ -n "$DB_URL" ] && [ -n "$DB_USERNAME" ] && [ -n "$DB_PASSWORD" ]; then
   echo "所有環境變數已設置，生成配置文件..."
   ./config.sh $CONFIG_FILE
 else
@@ -13,7 +13,20 @@ else
 fi
 
 
-cd $PLAY_DIR
-cat $PLAY_DIR/RUNNING_PID | xargs kill -9
-rm $PLAY_DIR/RUNNING_PID
-exec $PLAY_DIR/bin/taitra -Dfile.encoding=UTF-8 -Duser.timezone=Asia/Taipei -Dhttp.port=9090 -Dconfig.file=$CONFIG_FILE -Dlogger.file=$PLAY_DIR/conf/logback.xml >> /dev/null &
+
+# Safely kill the process if RUNNING_PID exists and contains a valid PID
+if [ -f "$PLAY_DIR/RUNNING_PID" ]; then
+  cat $PLAY_DIR/RUNNING_PID | xargs kill -9
+  rm $PLAY_DIR/RUNNING_PID
+else
+  echo "RUNNING_PID 文件不存在，無需刪除。"
+fi
+
+# Start the Play application
+exec $PLAY_DIR/bin/taitra \
+  -Dfile.encoding=UTF-8 \
+  -Duser.timezone=Asia/Taipei \
+  -Dhttp.port=9090 \
+  -Dconfig.file=$CONFIG_FILE \
+  -Dlogger.file=$PLAY_DIR/conf/logback.xml \
+  >> /dev/null 2>&1 &
